Add design doc for panics and errors.
Add a design doc that proposes a set of conventions and trade offs
involved in deciding whether to use a panic or an error in the
framework. The goal of this document is to help us achieve consistency
on this, and get the most out of Go's tools around exceptional behavior.

This proposal can be summarised as "use whichever is most appropriate
for the situation you're in", but outlines a number of different
guidelines to use and things to consider when determining which is most
appropriate.