Add design doc about the factory problem.
We want to instantiate new instances of resources, data sources, and
providers so that we can have some isolation at runtime, but either the
provider defined the type of the resource/data source/provider and we
don't know how to create new instance of it, or the framework defined
the type and the provider needs to supply values for it, which the
framework needs to get from the provider when creating a new instance of
it.

This document explores different ways to achieve that.