Start defining the Schema type.
Start building out the schema type, leaving some TODOs for things I'm
still mulling over and thinking about.

This also brings in the StringKind type, which is a pattern I'm not
thrilled about and we should probably think about it some more.

I also included the AttributeType and AttributeValue types, which I'll
add a package with some builtins for in a future commit, to show off how
they're used.