Handle null and unknown values.
Make it configurable whether null and unknown values can be substituted
for empty values when they're not supported by the type they're being
put into.

Make it so types can have a SetUnknown and SetNull method that's called
to handle null and unknown values.

Let pointers to pointers, pointers to maps, pointers to slices, and
pointers to interfaces be set to nil when there's a null value.

Fix a potential panic in reflectPrimitive by making sure the value can
be set before we try to set it.