Test helpers, and fix some bugs/simplify.
Add test for reflection helpers.

Simplify canBeNil. It used to check that a type can hold a nil value
_and_ can be set, but we already have a method to check if something can
be set: reflect.Value.CanSet. So we don't need to mirror that, which
lets us simplify to just "is this of a type that can be set to nil?"
This also fixes bugs in the canBeNil implementation, which wouldn't
consider things like slices on a struct pointer to be nillable.

Also, fix some bugs in setting things to zero; we were wrapping
reflect.Values in another reflect.Value (oops) and we weren't checking
if dereferencing the pointer would let us set it, which seems like a
nice thing to do.