Text objects, fix slices, simplify primitives.
Add some more tests for object to struct conversion.

Remove a log line. Oops.

Simplify primitives by passing a concrete type to tftypes.Value.As and
then setting a reflect.Value from the result. This avoids shenanigans
around pointers on reflect.Value types.

Make elements of a slice always settable by adding the reflect.Value to
the slice, then retrieving it by index, before recursing into it with
the `into` function. This means even when we add a new reflect.Value to
a slice, we can still set its value.

Add TODOs for the object tests we should probably write and the fact
that we've just ignored maps entirely so far.