Add support for maps, and change paradigm.
Add support for maps. Maps, of course, have to be an entirely
unaddressable type, meaning we can't use `Set` on the values of a map.
This is incredibly annoying and basically means all the code we had
already written wouldn't work for map values.

That's obviously not going to work, so I changed the paradigm a bit.
Instead of passing around a reflect.Value that these functions are
setting values on, these functions are now _constructing_ reflect.Values
of the same type as the reflect.Value they're given, filled with the
data from the tftypes.Value. This means we don't need to worry about the
Set semantics anymore. We'll build up a Value, then the entrypoint Into
function can set that Value on whatever is being passed in.

Also, instead of using trueReflectValue everywhere to let us ignore
pointers, the fact that we're constructing reflect.Values means we need
to actually care about pointers and interfaces, because you can't set a
string on a *string, they're different types. So we need to handle the
interfaces and pointers explicitly now, which I've left TODO comments
for.

But hey, the tests all pass, so, that's cool.