Add support for pointers and interfaces.
Add support for our setUnknownable and setNullable interfaces, as well
as for the AttributeValue interface and the tftypes.ValueConverter
interface.

This involved some sleight of hand. We want to have an empty version of
the type implementing the interface, but that type may be a pointer, so
an empty version would be nil, which leads to accidental panics. Which
we don't want. Worse, it could, in theory, be a pointer to a pointer to
a pointer to a pointer to a pointer to a pointer... you get the idea. So
we needed a new `pointerSafeZeroValue` that would find whatever type
underlies all the pointers, create a zero value of it, and then create a
pointer to that zero value, matching the level of indirection.

We also stopped converting values to interfaces to call the method
directly, letting us avoid nil pointer panics. It's more verbose, but
using reflection lets us have zero values and just not worry about
things that are nil. Much better.

Add support for pointers, dereferencing them and reflecting on the type
they implement, then wrapping that in a pointer on the way back out.
This also involves some sleight of hand to make the pointer settable,
but with some ingenuity and some cursed convoluted code, we manage it.
See comments in the reflectPointer function for more details.

Don't handle interface kinds, because there's no solid type underlying
them, and I don't feel like picking one. People can set real types, it's
ok.

Add tests for setUnknownable, setNullable, and AttributeValue values, as
well as for maps and pointers.

Supporting AttributeValue as targets required AttributeValues to gain a
SetTerraformValue method, updating the contents to match a
tftypes.Value. This required our AttributeValues to become pointers.
Annoying. It also required our composite AttributeValues to have
AttributeTypes as the ElemType, so they could create new AttributeValues
in the AttributeValue.SetTerraformValue method, if needed. This, in
turn, required AttributeType to gain an Equal method, so the
AttributeValue Equal method could compare AttributeTypes. Exhausting.

Also, we're actually testing that complex structs are having their
properties set correctly now, which is novel and neat. (They are.)