tfsdk: Automatic detection of attribute value normalization
It's common for Terraform's very general definition of value equality to
be stricter than a remote system's idea of equality. In that case, part of
a provider's job is to classify each difference in value as either
normalization or drift, or in other words to decide if a particular change
is functionally significant to the remote system or if it's just a
different way to write down what is effectively the same value.

The attr.Value interface already includes a method Equal, which allows
defining value equality for a custom type in a more liberal way than
Terraform's type system would define it.

Now we'll use that Equal method to automatically classify attribute value
changes as normalization if the schema-defined type considers the old and
new value as equal. In practice, that means that both ReadResource and
PlanResourceChange will retain an attribute value from the prior state
if the replacement new value is non-equal in Terraform terms but equal by
the provider's own definition.

Terraform Core's resource instance change lifecycle rules explicitly call
for providers to return the prior state value if the new value from
configuration is functionally-equivalent during PlanResourceChange. The
documented rules don't discuss the expected behavior of ReadResource at
all, but ReadResource also ultimately contributes to a diff (between the
prior state and the refreshed state) and it's important to use the same
classification rules in both cases so that refreshing doesn't create
"permadiffs" or situations where it takes multiple runs of
"terraform apply" before the configuration and state can fully converge.