types: SpecializedStringType prototype
It seems to be a common case that an API expects a string but imposes some
extra normalization and validation rules on that string, beyond
Terraform's very general idea of what a string is.

This is a prototype of a relatively easy-to-use helper to represent cases
like that. It behaves largely the same as types.String, but allows for
a custom validation rule and a custom normalization function, where the
normalization function changes the behavior of Value.Equal to consider
two values as equal if they have the same normalized form.