diag: Merge Diagnostic interface types, add SetDetail(), SetPath(), SetSeverity(), SetSummary() methods, and consolidate concrete types into GenericDiagnostic
Reference: https://github.com/hashicorp/terraform-plugin-framework/issues/169

Attribute path information is a fundamental part of diagnostics that get sent across the protocol. In some cases, called logic may not know the attribute path in which the calling logic must add it. Rather than treating diagnostics without paths as a separate interface than diagnostics with paths, this consolidates the two interfaces.

Setter methods are added to the `Diagnostic` interface type to support the use case mentioned above, as well as other logic which may want to modify diagnostic fields to improve practitioner and provider developer experience with the messaging or severity.