internal: Disable staticcheck deprecation warnings on migrated code
Since the code is now external to the package that declares these as deprecated, staticcheck will raise an issue. This silences those reports, either by denoting that the support is required in the case of Block support, or denoting that the functionality will be removed in the case of ResourceImportStateNotImplemented().