internal: Migrate ValidateProviderConfig RPC to fromproto6, fwserver, and toproto6
Reference: https://github.com/hashicorp/terraform-plugin-framework/issues/215

One particular quirk that this implementation shows is that it is not entirely possible to avoid working with `tfsdk`/`fwserver` bits in `proto6server` as the current data types within the framework (`Config`, `Plan`, `State`) require the `Schema` so data can be appropriately converted from the protocol to the framework. To support this effort and prevent repetitively executing the provider defined `GetSchema` method, a `fwserver` method is introduced that caches the `Schema` and `Diagnostics` results on first use. This methodology can also be applied in the future to resolve https://github.com/hashicorp/terraform-plugin-framework/issues/299.