tfsdk: Ensure Config, Plan, and State PathMatches will return null and unknown parent paths
This will also now return an error diagnostic if a given path expression is invalid for the schema data.

The `(path.Expression).MergeExpressions(...path.Expression) Expressions` is a helper method for the common use case of combining an attribute path expression with an incoming collection of expressions to perform generic validation across attributes.