website: Create concept-specific configuration validation pages
In an effort to show provider developers the per-concept functionality, dedicated pages have been created.

This also moves the current providers page as an index of a new Providers subsection of the navigation, which will likely contain other pages.

Lastly, this adjusts the navigation wording of resource pages for active voice consistency.