Shortcut logging when the level isn't active
This commit addresses this issue:
https://github.com/hashicorp/terraform-plugin-framework/issues/721

When working with large terraform resource attributes, logging can
become the primary cost to running plan/apply. This is because the
context logger used by tfsdklog is very expensive, the SDK framework
itself is very verbose, and even the log level means Debug/Trace logs
won't be emitted we still build the log.

As we log via the internal logging package anyway, we can avoid this
performance issue by maintaining a package-local log-level which is
initialised at the same time as we init the contextual logger. Then our
logging package can avoid calling into the tfsdklog (which is where we
pay the performance penalty) unless it believes the level is active,
which in most cases avoids 99% of the log cost.