internal/fwserver: Remove old multiple pass code comments in PlanResourceChange logic
Reference: https://github.com/hashicorp/terraform-plugin-framework/issues/183

When plan modification logic was introduced into the framework, it was not clear whether a multiple pass approach to running provider logic would be necessary. After thorough investigation, the multiple pass approach was instead replaced with explicit attribute default functionality to cover the main use case for such an approach, but the code comments were not removed when that was completed. To prevent future maintainer confusion, this is removing the old code comments.